+++ Platipus.Wallet.Api/Application/Responses/Vegangster/Base/VegangsterCommonResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace Platipus.Wallet.Api.Application.Responses.Vegangster.Base;
 
-public abstract record VegangsterCommonResponse(string Status, string Currency, int Balance);
+public abstract record VegangsterCommonResponse(string Status, string Currency, long Balance);
